Abstract validation to a separate file
